Add nagios IRC bot
Koumbit is using an IRC bot that Micah provided. It is a pair of perl
scripts that send Nagios notifications as messages in an IRC channel.

Add a class to make installing this IRC bot easy. It also defines
commands 'notify-by-irc' and 'host-notify-by-irc' that can be used with
checks to send notifications via the bot.

Signed-off-by: Gabriel Filion <lelutin@gmail.com>