Add missing Nagios resource types to base.pp
Some Nagios resource types are currently not being managed. This has the
result that exported resources for those types are not collected by
nagios servers, and also that configuration files for those resources
are put in their default path which is wrong for some OSes and with
possibly wrong permissions.

Add the missing resources to base.pp

Signed-off-by: Gabriel Filion <lelutin@gmail.com>