try to fix nagios::plugin problem
We have a serious problem with the concept that we had up to now if
you start adding more than one nagios server. This is an attempt to
fix it, with certain drawbacks, such as that you have to list all
the nagios plugins in the nagios modules. This will be subject of
further discussion.