Define check_load as a function of the nb of cpus
The "check_load" nagios check should not be hardcoded to some arbitrary
value. The thresholds should rather be adjusted in function of how many
CPUs are present in the server.

The factors were chosen by prior experience but could be adjusted if
needed.

Factors for warning thresholds (where N is the nb of cpus):
 * 1min : 5 * N
 * 5mins : 4 * N
 * 15mins : 3 * N

Factors for critical thresholds:
 * 1min : 10 * N
 * 5mins : 9 * N
 * 15mins : 8 * N