Add a nagios::nrpe::xinetd class
We don't really manage everything with this class (e.g. not the config
for running NRPE under xinetd). We ensure that the nrpe service is not
running since xinetd is the one that's supposed to spawn nrpe.

This way we can manage the NRPE config and commands, but it does not
conflict with xinetd.