IRC-bot: use 'enable' for Debian
Since the service "disable" bug was fixed in the debian service provider
[1] in the puppet 2.6.2-5 package in squeeze, let's switch from the
manual calls to the "enable" feature.

[1]: http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=573551

Also add a requirement between service[nagios-nsa] and
file[/etc/init.d/nagios-nsa]. This was ensured in Debian with the exec
that we're now replacing but didn't find its way through the
centosification.