Provide default config files for nagios.cfg and cgi.cfg
The current code tries to get the content of those two files in a bunch
of places and defines a fallback to a generic file in the module's
files/configs directory.

The problem is that there is no such generic config file present.

Link to Debian's config files as a generic version of those
configuration files.

This bug was found by taggart.